more refactoring for mvc - the controllers are the biggest parts missing
